move load and run into the main app file
